moved some functions into client.py; returns now 3hr data; cronjob script
