+++ WWCS/dashboard/service/get_open_meteo/client.py
@@ -1,7 +1,32 @@
+from datetime import datetime, timedelta
+import pandas as pd
+import mysql.connector
+import numpy as np
+from openmeteo_sdk.Variable import Variable
+
 import openmeteo_requests
 import requests_cache
 import retry_requests
 
+from common import USERNAME, PASSWORD
+
+def get_sites():
+    with mysql.connector.connect(
+        user=USERNAME,
+        password=PASSWORD,
+        host='127.0.0.1',
+        database='SitesHumans',
+    ) as cnx:
+        with cnx.cursor() as cursor:
+            cursor.execute("SELECT siteID, latitude, longitude FROM Sites WHERE siteID NOT LIKE '%-S%'")
+            return cursor.fetchall()
+
+def enum_code_to_name(enum_cls, code: int):
+    # enum members are stored as class attributes; reverse-lookup by value
+    for name, value in enum_cls.__dict__.items():
+        if value == code:
+            return name
+    return None
 
 class Client:
 
@@ -9,6 +34,74 @@ class Client:
         cache_session = requests_cache.CachedSession('.cache', expire_after=3600)
         retry_session = retry_requests.retry(cache_session, retries=5, backoff_factor=0.2)
         self.client = openmeteo_requests.Client(session=retry_session)
+
+    def _ensemble_response_to_dataframe(response):
+        """Convert ensemble response to DataFrame with mean and std."""
+        hourly = response.Hourly()
+
+        # Get time values
+        time_values = pd.date_range(
+            start=pd.to_datetime(hourly.Time(), unit="s", utc=True),
+            end=pd.to_datetime(hourly.TimeEnd(), unit="s", utc=True),
+            freq=pd.Timedelta(seconds=hourly.Interval()),
+            inclusive="left"
+        )
+
+        # Extract all temperature members
+        hourly_variables = [hourly.Variables(i) for i in range(hourly.VariablesLength())]
+        temp_vars = [v for v in hourly_variables
+                    if v.Variable() == Variable.temperature and v.Altitude() == 2]
+
+        # Stack all member values
+        member_values = []
+        for var in temp_vars:
+            #member = var.EnsembleMember()
+            values = var.ValuesAsNumpy()
+            member_values.append(values)
+
+        # Calculate mean and std across ensemble members
+        ensemble_array = np.stack(member_values)  # Shape: (members, time)
+        mean_values = np.mean(ensemble_array, axis=0)
+        std_values = np.std(ensemble_array, axis=0)
+
+        # Build DataFrame
+        df = pd.DataFrame({
+            'time': time_values,
+            'latitude': response.Latitude(),
+            'longitude': response.Longitude(),
+            'temperature_2m_mean': mean_values,
+            'temperature_2m_std': std_values,
+        })
+
+        # extract/return every 3rd hour        
+        return df[df["time"].dt.hour % 3 == 0]        
+
+    def _response_to_dataframe(response):
+        """Convert openmeteo-requests response to pandas DataFrame."""
+        hourly = response.Hourly()
+
+        time_values = pd.date_range(
+            start=pd.to_datetime(hourly.Time(), unit="s", utc=True),
+            end=pd.to_datetime(hourly.TimeEnd(), unit="s", utc=True),
+            freq=pd.Timedelta(seconds=hourly.Interval()),
+            inclusive="left"
+        )
+
+        data = {
+            'time': time_values,
+            'latitude': response.Latitude(),
+            'longitude': response.Longitude(),
+        }
+
+        for i in range(hourly.VariablesLength()):
+            var = hourly.Variables(i)
+            var_name = enum_code_to_name(Variable, var.Variable())
+            print(var_name)       
+            data[var_name] = var.ValuesAsNumpy()
+
+        # convert to panda and extract/return every 3rd hour
+        df = pd.DataFrame(data)
+        return df[df["time"].dt.hour % 3 == 0]
 
     def ensemble(self, params: dict):
         url = "https://ensemble-api.open-meteo.com/v1/ensemble"
@@ -24,3 +117,10 @@ class Client:
         response = responses[0] # ! this is correct since we use only one model/ensemble !
         return response
 
+    def ensemble_df(self, params: dict):
+        response = self.ensemble(params)
+        return _ensemble_response_to_dataframe(response)
+
+    def forecast_df(self, params: dict):
+        response = self.forecast(params)
+        return _response_to_dataframe(response)
