+++ WWCS/dashboard/service/get_open_meteo/get_open_meteo.py
@@ -1,104 +1,17 @@
-from datetime import datetime, timedelta
 from pathlib import Path
+import netCDF4
 
+from datetime import datetime, timedelta
 import pandas as pd
 import mysql.connector
-import netCDF4
 import numpy as np
 from openmeteo_sdk.Variable import Variable
 
-from client import Client
-from common import USERNAME, PASSWORD
+import client
 
 
-client = Client()
 
-def get_sites():
-    with mysql.connector.connect(
-        user=USERNAME,
-        password=PASSWORD,
-        host='127.0.0.1',
-        database='SitesHumans',
-    ) as cnx:
-        with cnx.cursor() as cursor:
-            cursor.execute("SELECT siteID, latitude, longitude FROM Sites WHERE siteID NOT LIKE '%-S%'")
-            return cursor.fetchall()
-
-def ensemble_df(params: dict) -> pd.DataFrame:
-    """Get ensemble data and return DataFrame with mean and std."""
-    response = client.ensemble(params)
-    return _ensemble_response_to_dataframe(response)
-
-def forecast_df(params: dict) -> pd.DataFrame:
-    response = client.forecast(params)
-    return _response_to_dataframe(response)
-
-def _ensemble_response_to_dataframe(response) -> pd.DataFrame:
-    """Convert ensemble response to DataFrame with mean and std."""
-    hourly = response.Hourly()
-
-    # Get time values
-    time_values = pd.date_range(
-        start=pd.to_datetime(hourly.Time(), unit="s", utc=True),
-        end=pd.to_datetime(hourly.TimeEnd(), unit="s", utc=True),
-        freq=pd.Timedelta(seconds=hourly.Interval()),
-        inclusive="left"
-    )
-
-    # Extract all temperature members
-    hourly_variables = [hourly.Variables(i) for i in range(hourly.VariablesLength())]
-    temp_vars = [v for v in hourly_variables
-                 if v.Variable() == Variable.temperature and v.Altitude() == 2]
-
-    # Stack all member values
-    member_values = []
-    for var in temp_vars:
-        #member = var.EnsembleMember()
-        values = var.ValuesAsNumpy()
-        member_values.append(values)
-
-    # Calculate mean and std across ensemble members
-    ensemble_array = np.stack(member_values)  # Shape: (members, time)
-    mean_values = np.mean(ensemble_array, axis=0)
-    std_values = np.std(ensemble_array, axis=0)
-
-    # Build DataFrame
-    df = pd.DataFrame({
-        'time': time_values,
-        'latitude': response.Latitude(),
-        'longitude': response.Longitude(),
-        'temperature_2m_mean': mean_values,
-        'temperature_2m_std': std_values,
-    })
-
-    return df
-
-def _response_to_dataframe(response) -> pd.DataFrame:
-    """Convert openmeteo-requests response to pandas DataFrame."""
-    hourly = response.Hourly()
-
-    time_values = pd.date_range(
-        start=pd.to_datetime(hourly.Time(), unit="s", utc=True),
-        end=pd.to_datetime(hourly.TimeEnd(), unit="s", utc=True),
-        freq=pd.Timedelta(seconds=hourly.Interval()),
-        inclusive="left"
-    )
-
-    data = {
-        'time': time_values,
-        'latitude': response.Latitude(),
-        'longitude': response.Longitude(),
-    }
-
-    for i in range(hourly.VariablesLength()):
-        var = hourly.Variables(i)
-        var_name = f"variable_{i}"
-        if i == 0:
-            var_name = 'temperature_2m'
-        data[var_name] = var.ValuesAsNumpy()
-
-    return pd.DataFrame(data)
-
+om_client = client.Client()
 
 def dataframe_to_netcdf(df: pd.DataFrame, filename: str, date_str: str):
     """
@@ -170,7 +83,7 @@ if __name__ == '__main__':
     today = datetime.today().date()
     dates = [d.strftime("%Y-%m-%d") for d in pd.date_range(today - timedelta(days=total_days), today)]
 
-    sites = get_sites()
+    sites = client.get_sites()
 
     outdir = Path('ifsdata')
     outdir.mkdir(exist_ok=True)
@@ -190,7 +103,7 @@ if __name__ == '__main__':
                 continue
 
             # Use ensemble API instead of forecast
-            df = ensemble_df({
+            df = om_client.ensemble_df({
                 'latitude': lat,
                 'longitude': lon,
                 'start_date': date_str,
